Compatible with RocksDB v6.16, #203
