+++ .github/workflows/ci.yml
@@ -1,0 +1,67 @@
+name: Continuous integration
+
+on: [push, pull_request]
+
+defaults:
+  run:
+    shell: bash
+
+jobs:
+  test:
+    strategy:
+      matrix:
+        os: [ubuntu-20.04, ubuntu-18.04, macos-11.0, macos-10.15]
+        go: [1.16.x, 1.15.x, 1.14.x, 1.13.x]
+    name: Go ${{ matrix.go }} tests @ ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Install Linux dependencies
+        if: startsWith(matrix.os, 'ubuntu-')
+        run: |
+          sudo apt-get update
+          sudo apt-get install -yq librocksdb-dev
+
+      - name: Install MacOS dependencies
+        if: startsWith(matrix.os, 'macos-')
+        run: |
+          brew install rocksdb
+
+      - name: Install Golang ${{ matrix.go }}
+        id: install-golang
+        uses: actions/setup-go@v2
+        with:
+          go-version: ${{ matrix.go }}
+      - run: go version
+
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Cache Golang modules
+        uses: actions/cache@v2
+        with:
+          path: ~/go/pkg/mod
+          key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
+          restore-keys: |
+            ${{ runner.os }}-go-
+
+      - name: Test v5
+        if: startsWith(matrix.os, 'ubuntu-')
+        run: go test -v
+
+      - name: Test v6.16 or later
+        if: startsWith(matrix.os, 'macos-')
+        run: go test -v -tags rocksdb_6_16
+
+  golangci:
+    name: lint
+    runs-on: ubuntu-20.04
+    steps:
+      - name: Install Linux dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -yq librocksdb-dev
+      - uses: actions/checkout@v2
+      - name: golangci-lint
+        uses: golangci/golangci-lint-action@v2
+        with:
+          version: v1.38
