+++ db.go
@@ -577,89 +577,6 @@ func (db *DB) DropColumnFamily(c *ColumnFamilyHandle) error {
 	return nil
 }
 
-// GetApproximateSizes returns the approximate number of bytes of file system
-// space used by one or more key ranges.
-//
-// The keys counted will begin at Range.Start and end on the key before
-// Range.Limit.
-func (db *DB) GetApproximateSizes(ranges []Range) []uint64 {
-	sizes := make([]uint64, len(ranges))
-	if len(ranges) == 0 {
-		return sizes
-	}
-
-	cStarts := make([]*C.char, len(ranges))
-	cLimits := make([]*C.char, len(ranges))
-	cStartLens := make([]C.size_t, len(ranges))
-	cLimitLens := make([]C.size_t, len(ranges))
-	for i, r := range ranges {
-		cStarts[i] = (*C.char)(C.CBytes(r.Start))
-		cStartLens[i] = C.size_t(len(r.Start))
-		cLimits[i] = (*C.char)(C.CBytes(r.Limit))
-		cLimitLens[i] = C.size_t(len(r.Limit))
-	}
-
-	defer func() {
-		for i := range ranges {
-			C.free(unsafe.Pointer(cStarts[i]))
-			C.free(unsafe.Pointer(cLimits[i]))
-		}
-	}()
-
-	C.rocksdb_approximate_sizes(
-		db.c,
-		C.int(len(ranges)),
-		&cStarts[0],
-		&cStartLens[0],
-		&cLimits[0],
-		&cLimitLens[0],
-		(*C.uint64_t)(&sizes[0]))
-
-	return sizes
-}
-
-// GetApproximateSizesCF returns the approximate number of bytes of file system
-// space used by one or more key ranges in the column family.
-//
-// The keys counted will begin at Range.Start and end on the key before
-// Range.Limit.
-func (db *DB) GetApproximateSizesCF(cf *ColumnFamilyHandle, ranges []Range) []uint64 {
-	sizes := make([]uint64, len(ranges))
-	if len(ranges) == 0 {
-		return sizes
-	}
-
-	cStarts := make([]*C.char, len(ranges))
-	cLimits := make([]*C.char, len(ranges))
-	cStartLens := make([]C.size_t, len(ranges))
-	cLimitLens := make([]C.size_t, len(ranges))
-	for i, r := range ranges {
-		cStarts[i] = (*C.char)(C.CBytes(r.Start))
-		cStartLens[i] = C.size_t(len(r.Start))
-		cLimits[i] = (*C.char)(C.CBytes(r.Limit))
-		cLimitLens[i] = C.size_t(len(r.Limit))
-	}
-
-	defer func() {
-		for i := range ranges {
-			C.free(unsafe.Pointer(cStarts[i]))
-			C.free(unsafe.Pointer(cLimits[i]))
-		}
-	}()
-
-	C.rocksdb_approximate_sizes_cf(
-		db.c,
-		cf.c,
-		C.int(len(ranges)),
-		&cStarts[0],
-		&cStartLens[0],
-		&cLimits[0],
-		&cLimitLens[0],
-		(*C.uint64_t)(&sizes[0]))
-
-	return sizes
-}
-
 // SetOptions dynamically changes options through the SetOptions API.
 func (db *DB) SetOptions(keys, values []string) error {
 	num_keys := len(keys)
